api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
Pillow==10.3.0
